// ecis-nodejs-yseq-feat_soap1218/service/ecissdk/model/data.ts
export interface Headers {
  [key: string]: any
}
// ...
export interface OpenGetIdConfuseParams {
  queryStr: string // 格式：a=x&b=y
}

export interface OpenGetIdConfuseResp {
  result: string
  msg?: string
  data?: {
    confuse_result: string
  }
}

export interface OpenGetIdDeConfuseResp {
  result: string
  msg?: string
  data?: {
    de_confuse_result: {}
  }
}

export interface OpenGetFileInfoParam {
  queryStr: string // 格式：a=x&b=y
  headers: Headers
}

export interface OpenGetFileInfoResp {
  result: string
  msg?: string
  data?: {
    file: {
      id: string
      name: string
      version: number
      size: number
      creator: string
      create_time: number
      modifier: string
      modify_time: number
      download_url: string
      preview_pages: number
      user_acl: {
        rename: number
        history: number
        copy: number
        export: number
        print: number
      }
      watermark: {
        type: number
        value: string
        fillstyle: string
        font: string
        rotate: number
        horizontal: number
        vertical: number
      }
    }
    user: {
      id: string
      name: string
      permission: string
      avatar_url: string
    }
  }
}

export interface OpenPreviewInnerFileInfoParam {
  queryStr: string // 格式：a=x&b=y
  headers: Headers
}

export interface OpenPreviewInnerFileInfoResp {
  result: string
  msg?: string
  data?: {
    file: {
      id: string
      name: string
      version: number
      size: number
      creator: string
      create_time: number
      modifier: string
      modify_time: number
      download_url: string
      preview_pages: number
      user_acl: {
        rename: number
        history: number
        copy: number
        export: number
        print: number
      }
      watermark: {
        type: number
        value: string
        fillstyle: string
        font: string
        rotate: number
        horizontal: number
        vertical: number
      }
    }
    user: {
      id: string
      name: string
      permission: string
      avatar_url: string
    }
  }
}

export interface OpenGetSaveFileUrlParam {
  queryStr: string // 格式：a=x&b=y
  headers: Headers
}

export interface OpenGetSaveFileUrlResp {
  result: string
  msg?: string
  data?: {
    url: string
    params: any
    headers: any
  }
}